+++ app/src/main/java/com/altabel/coolmovies/entity/Movie.kt
@@ -4,7 +4,7 @@ import java.util.*
 
 data class Movie(
     val id: Int,
-    val popularity: Int,
+    val popularity: Float,
     val voteCount: Int,
     val video: Boolean,
     val posterPath: String,
@@ -14,7 +14,7 @@ data class Movie(
     val originalTitle: String,
     val genreIds: List<Int>,
     val title: String,
-    val voteAverage: Int,
+    val voteAverage: Float,
     val overview: String,
     val releaseDate: Date
 )
